requirements.txt
docopt
numpy
pillow

